Deprecate exporterhelperprofiles module in favor of xexporterhelper
to allow adding more experimental data types.
