+++ exporter/exporterhelper/xexporterhelper/profiles_batch_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package exporterhelperprofiles
+package xexporterhelper
 
 import (
 	"context"
